Note relative option numbers in parse output, use that info in sample code
The Makefile is smarter now, it's has better dependency chains and that
fixed a bug where the library wasn't updated if the ".c" code changed.

The parse_command_options routine now includes a field noting which
word number in triggered the parsing of that particular option.  The
point of that is to allow alternate forms of an option to map to the
same option number, and to let the calling program figure out which
to use if multiple flags are given on the command line.  Also, the
"flags" field is initialized differently and a bug in the code which
stores a linked list of unparsed arguments was fixed.