Increase SSL I/O buffer and fix misinterpreted SSL option value
Setting SSL_OP_MICROSOFT_BIG_SSLV3_BUFFER, despite the name, is
a just a way to get the underlying code to use a bigger I/O
buffer.  That should help get more accurate packet level perf
data.

Also, the parameter passed to SSL_CTX_set_read_ahead() is just
an on/off switch not a buffer size.  So change the constant to
"1" to try to clarify it's use.