Smarter "invalid data" handling.
Each x/y column is treated independently when scanning input records
in terms of whether or not to include the data now.  That allows for
multi-line charts where the only data plots for a given series is the
data that changes when two or more series share X or Y data but don't
always have the same number of cases.