Change the way matte size and font size are calculated.
The matte size wasn't leaving room for the circle outline and that
led to datapoints right on the edge of the chart being cutoff.

The font size picked was based solely on the width of the chart.
When height was adjusted to over 1/2 the width that caused the text
to be too wide.