Add the bulk of the code needed to enable "legends" in charts
The code to support legends describing all the data series on the
chart has been added to the core SVG library routines.  And the
basic pieces needed to enabled it in the "SVG-It" application are
done as well.

There's currently know way to set the descriptions of the data
series though.  So it's not completely done.