Add base mask layer covering the entire chart to avoid artifacts
If the chart is resized to only cover part of the reserved space
on the page, in order to preserve the x:y ratio, then the partially
transparent mask layer created shadow artifacts on the sides that
are not flush to the edge of the window.  This change adds a 100%
opaque mask layer under all the lines, circles, text, grids, etc...
to cover those up.