Quick fix to get SSL CA CERT's working on MacOS
One MacOS the trusted CA certs are not made available in /etc/ssl/certs
like they are on Linux.  Instead packages that aren't setup to use the
"keychain" rely on copies installed in alternate locations.  This change
tries one of the typical locations if compiled on MacOS.

A better solution should be designed at some point, since there's no
standard location for CA certs on MacOS and this won't work for everyone.